Move mainview to views directory
Refs #102
