+++ mslice/tests/powder_projection_presenter_test.py
@@ -1,10 +1,10 @@
 import mock
 import unittest
 
-from mslice.mainview import MainView
 from mslice.models.projection.powder.projection_calculator import ProjectionCalculator
 from mslice.presenters.interfaces.main_presenter import MainPresenterInterface
 from mslice.presenters.powder_projection_presenter import PowderProjectionPresenter
+from mslice.views.mainview import MainView
 from mslice.views.powder_projection_view import PowderView
 from mslice.widgets.projection.powder.command import Command
