+++ mslice/views/mainview.py
